// src/app/api/model/serviceupdateInfo.ts
/**
 * SalonMaster API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServiceupdateInfo { 
    nom?: string;
    description?: string;
    prix?: number;
    duree?: number;
}